+++ build.proj
@@ -1,5 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="15.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+<Project ToolsVersion="15.0" DefaultTargets="Build" InitialTargets="PrepareEnvironment" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <UsingTask AssemblyFile="$(TaskDirectory)Microsoft.DotNet.SourceBuild.Tasks/bin/Debug/netstandard1.4/Microsoft.DotNet.SourceBuild.Tasks.dll" TaskName="GetHostInformation" />
+
   <Import Project="dir.props" />
 
   <PropertyGroup>
@@ -21,10 +23,6 @@
     </CleanProjectTargets>
   </PropertyGroup>
 
-  <Target Name="Test">
-    <Message Importance="High" Text="$(Configuration) $(TargetOS) $(Platform)" />
-  </Target>
-
   <Target Name="Build" DependsOnTargets="PrepareOutput;$(BuildProjectTargets)">
     <Message Importance="Normal" Text="Build Complete" />
   </Target>
@@ -44,6 +42,18 @@
   <Target Name="CleanOutput">
     <RemoveDir Directories="$(OutputPath)" />
     <RemoveDir Directories="$(IntermediatePath)" />
+  </Target>
+
+  <Target Name="PrepareEnvironment">
+    <GetHostInformation Condition="'$(TargetRid)' == ''">
+      <Output PropertyName="TargetRid" TaskParameter="Rid" />
+    </GetHostInformation>
+
+    <GetHostInformation Condition="'$(TargetOS)' == ''">
+      <Output PropertyName="TargetOS" TaskParameter="OSName" />
+    </GetHostInformation>
+
+    <Message Text="Build Environment: $(Platform) $(Configuration) $(TargetOS) $(TargetRid)" />
   </Target>
 
   <Import Project="$(MSBuildThisfileDirectory)targets/build-standard.targets" />
