+++ dir.props
@@ -3,8 +3,6 @@
   <PropertyGroup>
     <Configuration Condition="$(Configuration) == ''">Release</Configuration>
     <Platform Condition="'$(Platform)' == ''">x64</Platform>
-    <TargetOS Condition="'$(TargetOS)' == '' and '$(OS)' == 'Windows_NT'">Windows_NT</TargetOS>
-    <TargetOS Condition="'$(TargetOS)' == '' and '$(OS)' != 'Windows_NT'">Linux</TargetOS>
   </PropertyGroup>
 
   <PropertyGroup>
